Add more definitions for instructions and tests for Validator
Combine SimpleValidator and Validator into a single file
Add more tests for branching and loops
